Trying to fix bugs

+++ controllers/bot_controller.rb
@@ -1,68 +1,74 @@
+require 'pry-byebug'
+
 class BotController
   def initialize(bot)
     @bot = bot
   end
 
   def redd_command(event)
-    user = set_user(event)
-    # channel = Channel.find_by_discord_id(event.channel.id)
-    return "Sorry you already have an active post, delete that first with d!delete before creating a new one." if user.active_post
-    return "Hmm. Please check if your post matches the template in #art-announcements or type d!template" unless event.content.include?("1.")
-    return "Sorry you're in the queue for a different post, please finish that before you enter a queue!" if user.in_queue
+      user = set_user(event)
+      # channel = Channel.find_by_discord_id(event.channel.id)
+      return "Sorry you already have an active post, delete that first with d!delete before creating a new one." if user.active_post
+      return "Hmm. Please check if your post matches the template in #art-announcements or type d!template" unless event.content.include?("1.")
+      return "Sorry you're in the queue for a different post, please finish that before you enter a queue!" if user.in_queue
 
-    announcement_message = Announcement.create!(content: event.content, user: user)
-    user.active_post = true
-    user.save!
-    art_pieces = announcement_message.extract_art_pieces
-    art_pieces.each_with_index do |art,i|
-      ArtPiece.create!(name: art, number: i + 1, status: "open", announcement: announcement_message)
-    end
-    bot_message = @bot.send_message(ENV['CHANNEL_ID'], "", nil, { description: announcement_message.original_message_no_art, fields: announcement_message.build_inline_fields, color: 0x12457E } )
-    announcement_message.discord_id = bot_message.id
-    announcement_message.save!
+      announcement_message = Announcement.create!(content: event.content, user: user)
+      user.active_post = true
+      user.save!
+      art_pieces = announcement_message.extract_art_pieces
+      art_pieces.each_with_index do |art,i|
+        ArtPiece.create!(name: art, number: i + 1, status: "open", announcement: announcement_message)
+      end
+      bot_message = @bot.send_message(ENV['CHANNEL_ID'], "", nil, { description: announcement_message.original_message_no_art, fields: announcement_message.build_inline_fields, color: 0x12457E } )
+      announcement_message.discord_id = bot_message.id
+      announcement_message.save!
 
-    add_reaction_event_listener(bot_message)
-    remove_reaction_event_listener(bot_message)
-    add_edit_event_listener(bot_message)
+      add_reaction_event_listener(bot_message)
+      remove_reaction_event_listener(bot_message)
+      add_edit_event_listener(bot_message)
 
-    @bot.send_message(event.channel.id, "", nil, { description: "Succesfully created! Check out #art-announcements. Use d!queue <dodo_code> to activate the post!", color: 0x12457E } )
+      @bot.send_message(event.channel.id, "", nil, { description: "Succesfully created! Check out #art-announcements. Use d!queue <dodo_code> to activate the post!", color: 0x12457E } )
 
-    nil
+      nil
   end
 
   def queue_command(event)
-    channel = @bot.channel(ENV['CHANNEL_ID'])
-    dodo = event.content.downcase.gsub("d!queue","").strip.match(/\w{5}/).to_s.upcase
-    event.message.delete
-    if dodo.empty?
-      @bot.send_message(event.channel.id, "", nil, { description: "Oops, I think your forgot the dodo code, or maybe you typed it wrong? Try again! Remember it's d!queue <dodo_code>. No brackets." , color: 0x12457E } )
-      return nil
-    end
-    user = User.find_by_discord_id(event.user.id)
-    announcement_message = Announcement.where(user: user).last
-    if announcement_message.dodo
-      claimed_art_pieces = announcement_message.art_pieces.where(status: "claimed")
-      user_ids = claimed_art_pieces.map(&:user).map(&:discord_id)
-      claimed_users = channel.users.select { |user| user_ids.any? { |id| id == user.id } }
-      claimed_users.each do |claimed_user|
-        claimed_user.pm("Sorry about the inconvenience. #{user.mention}'s island is back up. Here's the new dodo code: #{dodo}")
+      channel = @bot.channel(ENV['CHANNEL_ID'])
+      dodo = event.content.downcase.gsub("d!queue","").strip.match(/\w{5}/).to_s.upcase
+      event.message.delete
+      if dodo.empty?
+        @bot.send_message(event.channel.id, "", nil, { description: "Oops, I think your forgot the dodo code, or maybe you typed it wrong? Try again! Remember it's d!queue <dodo_code>. No brackets." , color: 0x12457E } )
+        return nil
       end
-      @bot.send_message(event.channel.id, "", nil, { description: "Dodo code updated!", color: 0x12457E } )
-    else
-      @bot.send_message(event.channel.id, "", nil, { description: "Gotcha! Your post is now active!", color: 0x12457E } )
-    end
-    announcement_message.dodo = dodo
-    announcement_message.save!
-    bot_message = channel.load_message(announcement_message.discord_id)
-    EMOJIS.first(announcement_message.art_pieces.count).each do |emoji|
-      bot_message.react emoji
-    end
-    nil
+      user = User.find_by_discord_id(event.user.id)
+      announcement_message = Announcement.where(user: user).last
+      if announcement_message.dodo
+        claimed_art_pieces = announcement_message.art_pieces.where(status: "claimed")
+        user_ids = claimed_art_pieces.map(&:user).map(&:discord_id)
+        claimed_users = channel.users.select { |user| user_ids.any? { |id| id == user.id } }
+        claimed_users.each do |claimed_user|
+          claimed_user.pm("Sorry about the inconvenience. #{user.mention}'s island is back up. Here's the new dodo code: #{dodo}")
+        end
+        @bot.send_message(event.channel.id, "", nil, { description: "Dodo code updated!", color: 0x12457E } )
+      else
+        @bot.send_message(event.channel.id, "", nil, { description: "Gotcha! Your post is now active!", color: 0x12457E } )
+      end
+      announcement_message.dodo = dodo
+      announcement_message.save!
+      bot_message = channel.load_message(announcement_message.discord_id)
+      EMOJIS.first(announcement_message.art_pieces.count).each do |emoji|
+        bot_message.react emoji
+      end
+      nil
   end
 
   def buy_command(event)
     channel = @bot.channel(ENV['CHANNEL_ID'])
     user = User.find_by_discord_id(event.user.id)
+    if !user.in_queue
+      @bot.send_message(event.channel.id, "", nil, { description: "Sorry! You haven't claimed any art", color: 0x12457E } )
+      nil
+    end
     bought_art = user.art_pieces.where(status: "claimed").first
     bought_art.status = "bought"
     bought_art.save!
@@ -128,17 +134,19 @@ class BotController
   end
 
   def remove_reaction_event_listener(bot_message)
-     EMOJIS.each_with_index do |emoji, i|
+    EMOJIS.each_with_index do |emoji, i|
       @bot.reaction_remove(attributes = { emoji: emoji }) do |event|
         user = set_user(event)
-        user.in_queue = false
-        user.save!
         reacted_message = Announcement.find_by_discord_id(event.message.id)
         art_piece = reacted_message.art_pieces.where(number: i + 1).first
         reaction = user.reactions.where(announcement: reacted_message).first
-        reaction.destroy! unless art_piece.status == "bought"
+        if reaction
+          reaction.destroy! unless art_piece.status == "bought"
+        end
         art_piece.update_status unless art_piece.status == "bought"
         bot_message.edit("", { description: reacted_message.original_message_no_art, fields: reacted_message.build_inline_fields })
+        user.in_queue = false
+        user.save!
       end
     end
   end
@@ -148,6 +156,7 @@ class BotController
       @bot.reaction_add(attributes = { emoji: emoji }) do |event|
         user = set_user(event)
         reacted_message = Announcement.find_by_discord_id(event.message.id)
+        binding.pry
         if user.in_queue
           event.user.pm("Ooops! You're already in a another queue.")
           bot_message.delete_reaction(user.discord_id, emoji)
@@ -155,12 +164,12 @@ class BotController
         end
         art_piece = reacted_message.art_pieces.where(number: i + 1).first
         if user.can_react?(art_piece)
-          user.in_queue = true
-          user.save!
           Reaction.create!(number: i + 1, announcement: reacted_message, user: user)
           art_piece.update_status
           bot_message.edit("", { description: reacted_message.original_message_no_art, fields: reacted_message.build_inline_fields })
           event.user.pm("Get ready pick up your art at #{reacted_message.user.mention}'s island! Dodo code is: #{reacted_message.dodo}")
+          user.in_queue = true
+          user.save!
         else
           bot_message.delete_reaction(user.discord_id, emoji)
           event.user.pm("Sorry! You can't claim this! If you don't know why, please ask a moderator!")
